// frontend/src/client/api/models/SearchTag.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SearchTag = {
  label: SearchTag.label
  text?: string
}
export namespace SearchTag {
  export enum label {
    OFFICIAL = 'official',
    VALID = 'valid',
    LATEST = 'latest',
    DOWNLOADED = 'downloaded',
    AUTHOR = 'author',
    TAG = 'tag'
  }
}