// frontend/src/client/api/models/Adapter.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ModuleTag } from './ModuleTag'
export type Adapter = {
  module_name: string
  project_link: string
  name: string
  desc: string
  author: string
  homepage: string
  tags?: Array<ModuleTag>
  is_official: boolean
  is_download?: boolean
}